feat: add shell script to launch docker environment

+++ currentaccounts/download-opentracing-specialagent.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
-wget -O opentracing-specialagent-1.7.4.jar "https://repo1.maven.org/maven2/io/opentracing/contrib/specialagent/opentracing-specialagent/1.7.4/opentracing-specialagent-1.7.4.jar"
+if [ ! -f ./opentracing-specialagent-1.7.4.jar ]; then
+    echo "File opentracing-specialagent-1.7.4.jar not found! and will be downloaded!"
+    wget -O opentracing-specialagent-1.7.4.jar "https://repo1.maven.org/maven2/io/opentracing/contrib/specialagent/opentracing-specialagent/1.7.4/opentracing-specialagent-1.7.4.jar"
+else
+    echo "File opentracing-specialagent-1.7.4.jar found! and no download is needed!"
+fi
